reset `customized` branch in theme/academic
untrack the updates in origin/master branch as there are some
significant versioning issue with the latest commits and `hugo server`
keeps getting errors.

use this version of teh academic theme to update the website. will think
about how to track and sync with the latest updates in the futin the future.